requirements.txt
numpy==2.2.3
pickle-mixin==1.0.2
streamlit==1.42.1
streamlit-option-menu==0.4.0
scikit-learn==1.6.1
streamlit-authenticator==0.2.2
PyYAML==6.0.1     # Required for loading the config.yaml
bcrypt==4.1.2     # Required for password hashing
